+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added item checkout and scan-in.
 - Analytics: The `getRows()` method is deprecated. Iterate over the `Report` instead.
 - Made package auto-discoverable in Laravel 5.5+.
+- Added helper method `Users::findOne()` to get the first match from a search.
 
 ### Changed